+++ src/main/java/next/controller/UserController.java
@@ -5,7 +5,7 @@ import core.annotation.web.RequestMapping;
 import core.annotation.web.RequestMethod;
 import core.annotation.web.RequestParam;
 import core.db.DataBase;
-import core.mvc.JspView;
+import core.mvc.view.JspView;
 import core.mvc.ModelAndView;
 import next.model.User;
 import org.slf4j.Logger;
